Improve image rendering
When the image view comes on screen, it checks if the frame size would benefit from downsampling. This is primarily relevant for full screen mode, where the view size is likely to be larger than the image itself. When the view leaves the screen, however (disappears), it removes its associated image from state, freeing up memory. This process is a bit aggressive, as it always does it, so in the future, I may only apply this when memory has reached a certain level.