Scope Live Text settings to scene more appropriately
The previous implementation disabled Live Text for the scene when it was disabled in the app settings, but would intentionally not enable it when it was re-enabled. It was more or less a crude implementation built around .onAppear/.onChange

This new implementation uses a Menu to toggle Live Text and the icons, and syncs with the app settings unless the scene value has explicitly been changed. It has a little bit of complexity embedded in it, since there's no way to get the setting synced again, but it's much nicer (especially since toggling is now immediately reflected).