Fix "Open in Sidebar" sometimes not opening
This also improves the UI for listing copy destinations. Before, only the folder name was presented, making it not easily possible to distinguish folders with the same name but different paths. This implementation takes the minimum paths required to distinguish a folder from others and presents them in a path-like structure.

For example, on my system, I have two folders:
- "~/Desktop/Anime & Manga/The Ancient Magus' Bride/Pages"
- "~/Desktop/Anime & Manga/Children of the Whales/Pages"

Previously, in the UI, the user would only see "Pages" for both of them, which was not helpful. Now they'll see "The Ancient Magus' Bride 􀰇 Pages" and "Children of the Whales 􀰇 Pages", which is a lot more useful.