Refactor security scope handling
Previously, we made assumptions about what URLs from the system were security scoped. This was cubersome to handle and did not yield much benefit, since a URL that wasn't security scoped in one version of macOS could be later on (looking at you, drag & drop). This reimplementation just tries to access the scope and, if it's not available, log a warning while continuing.

As for new user-facing features:
- Windowless mode allows users to see just the image (and not the window's other components, like the toolbar). This was a feature at the start of the development of the app that I removed since I found it difficult to implement with List.