Add support for opening folders
This took me ~3 days to implement. If there's one major takeaway from this, it's that security scoped bookmarks are super complex. Even this implementation, which is very careful about how to create and access them has minor bugs plaguing it. But it was totally worth it, and has sped up resolution even more (The Ancient Magus' Bride collection now takes under a second to load in *development build*!)

Some caveats:
- Hashable conformance to switch to the existing window if the set of images is the same seems to no longer work.
- I've commented out the code responsible for reloading an image if it fails to load due to it being inaccessible. I can easily add it back, but am unsure since I haven't been satisfied with the implementation.