Use anchor preferences instead of GeometryReader for determining visible items
The main star of this update is the replacement of GeometryReader for determining the images on screen for anchor preferences. It still relies on GeometryReader, but in an isolated region of the view hierarchy. This creates some massive wins for the user:
- No more hangs when loading new images
- Determining which images are on screen is more accurate, making the title and sidebar feel snapper (up to Preview's level)
- The setting about the potential performance degredation from tracking the current image is no longer relevant, and so has been removed.

Along with the change came some improvements:
- DisplayView no longer requires wrapping content in GeometryReader, which should improve performance
- SwiftUI is now able to release phases on its own, meaning the memory creep should be more under control
- Users can now hit Command-Shift-T to toggle Live Text highlighting

There is unfortunately a loss in this change: usage of focused (scene) values. For some reason, changes in large view hierarchies produces hangs that would be worse than before. As a result, jumping to the current image has been temporarily removed (until I find a solution, at least). Along with this, the Live Text highlighting feature is implemented via a hidden Button, which would be better done via a menu bar item.