Use immutable bookmarks architecture (cont.)
This is a continuation of the previous commit.

There are way too many changes to be noted in the commit, but the grand change is the restructuring of the model to factor bookmarks into their own structure (a bookmark store). Previously, the handling of bookmarks was coupled to the model, which made it difficult to reuse and require a lot of mutability. This new model attaches an identifier to concerned model items, in which they can lookup the bookmark in a bookmark store. The bookmark store is built on a layer of generic bookmark structures. In a way, the model could be thought of like so:

URL -> Bookmark -> Bookmark store

Bookmark store -> Image collection
Bookmark store -> Copy depot

And because bookmark is its own layer, image collection and copy depot can easily work on them to eventually apply to the bookmark store. The new architecture is massively simpler, and despite being immutable, is faster and more stable than my prior implementation.

For user-facing features, some have been temporarily disabled, such as automatically refreshing an image with a stale bookmark. However, for new features, "Hide the toolbar when scrolling" has been expanded into three options for hiding: the toolbar, the cursor, and the scroll bar. In addition, a bug where the "drop images here" button would slightly stay when loading a new collection has been resolved.