Generalize LiveTextView
Previously, LiveTextView was responsible for a lot of domain-specific logic (analyzing, retrying, caching, etc.) This new implementation makes it the instantiator's responsibility, so ImageCollectionDetailItemView is handling that process now. As a result, the user no longer needs to worry about Live Text failing and it trying to continuously analyze the image as an effect.

There are two new features regarding Live Text:
- A setting to display the "Search in ..." menu item (with a note that it'll always open in Safari)
- A setting to downsample images that are too large to be analyzed (with a note that it'll be written to disk)
- A new settings panel for the options (since putting them in General is a bit much)

In addition, some other changes have been made:
- Image collections will no longer reload when changes have been made to them
- A bug that caused the last image in a collection to not have a proper margin has been fixed