Improve sidebar
The sidebar now has a number of "states" it can be in (currently normal, bookmark, and search). This allows the sidebar to maintain a uniform but unique set of information, depending on the context the user is in.

Speaking of search, users can now search images by their name to quickly find matches. There were previous attempts at implementing this with SwiftUI's searchable functionality, but it never behaved well (often just crashing). Searching can be accessed via the sidebar (either the search button or invisible text field), which will display a subset of the images.

Scrolling the sidebar and main canvas is no longer dependent on the images in the collection, removing a performance constraint. A big part of this implementation was remodeling how scrollers are threaded through the view hiearchy. Rather than having the sidebar and detail views declare their focus values with the image collection threading it through views, the image collection picks up the focus values and places them in the environment. This is possible due to the scroller being primarily dependent on its input, as well as its identity containing a default case (to make sure SwiftUI updates appropriately).

Previously, one may have saw the "Drop images here" button for a split second before an image collection started loading; this has been fixed with a new loading indicator in the environment.