Add basic installation info
The version and build numbers have also been adjusted to reflect the state of the project. At the moment, we're on version 0.1.0 and don't use a build number to indicate the project is at least usable by others now. Managing build numbers, however, is a bit annoying and only really useful for automated tooling.