Replace back/forward navigation for current image restoration
The back/forward menu implemented via the app menu and toolbar was used to navigate images interacted with via the sidebar. When a user selected an image via the sidebar, its ID was recorded to support navigation. While it worked, it was limited in two ways: direction and rationale. For the former, the navigation menu had two options for ordering items: by interaction or by sorting in the sidebar. Advance chose the later, which naturally resulted in odd expressions (e.g. "Back" navigating to the closest image before the current, as opposed to the last interaction). For the latter, the ultimate purpose was to support navigation that survived the app lifecycle (that is, restoring the navigation state after relaunch). As a result, this often meant the last image by sidebar sort order was the most relevant item.

This commit simplifies the expression so restoring the last image automatically happens. It's much simpler, as it doesn't require direct interaction, and achieves a common want (resuming where one left off).